requirements.txt
fastapi
uvicorn
pydantic
transformers
torch
langchain
openai
python-dotenv
requests
